+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - `Opis\View\EngineEntry` class was moved into its own file
 - `Opis\View\ViewRouter` was renamed to `Opis\View\ViewApp`
+- Source files were moved from `lib` to `src`
 
 ### Fixed
 - Nothing